requirements.txt
asgiref==3.3.1
click==7.1.2
fastapi==0.63.0
googleapis-common-protos==1.52.0
grpcio==1.36.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.1.1
opentelemetry-api==1.0.0rc1
opentelemetry-exporter-jaeger==1.0.0rc1
opentelemetry-instrumentation==0.18b0
opentelemetry-instrumentation-asgi==0.18b1
opentelemetry-instrumentation-fastapi==0.18b1
opentelemetry-sdk==1.0.0rc1
opentelemetry-util-http==0.18b1
protobuf==3.15.7
pydantic==1.8.1
six==1.15.0
starlette==0.13.6
thrift==0.13.0
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
wrapt==1.12.1
